feat: show the available vacancies of a hotel
