+++ src/components/Accommodation/index.js
@@ -3,9 +3,7 @@ import styled from 'styled-components';
 
 import useHotel from '../../hooks/api/useHotel';
 import useTicket from '../../hooks/api/useTicket';
-
-import { HotelsWrapper } from './HotelsWrapper';
-import { SectionWrapper } from './SectionWrapper';
+import Hotels from './Hotels';
 
 export default function Accommodation() {
   const { ticket } = useTicket();
@@ -15,56 +13,11 @@ export default function Accommodation() {
     return ticket && ticket.status === 'PAID' && !ticket.TicketType.isRemote && ticket.TicketType.includesHotel;
   }
 
-  function getRoomCapacity({ roomsCapacity }) {
-    const capacity = {};
-
-    roomsCapacity = roomsCapacity.filter((number) => {
-      if (!capacity[number]) {
-        capacity[number] = true;
-        return true;
-      }
-      return false;
-    });
-
-    return roomsCapacity
-      .sort()
-      .join(', ')
-      .replace(1, 'Single')
-      .replace(2, 'Double')
-      .replace(3, 'Triple')
-      .replace(/,\s([^,]+)$/, ' e $1');
-  }
-
   return (
     <>
       <StyledTypography variant="h4">Escolha de hotel e quarto</StyledTypography>
 
-      {isValidTicket(ticket) && (
-        <SectionWrapper>
-          <h2>Primeiro, escolha seu hotel</h2>
-
-          <div>
-            {hotels ? (
-              hotels.map((hotel, index) => (
-                <HotelsWrapper key={index} isLast={index === hotels.length - 1}>
-                  <img alt={hotel.name} src={hotel.image} />
-                  <h4>{hotel.name}</h4>
-
-                  <span>
-                    <b>Tipos de acomodação: </b>
-
-                    <br />
-
-                    {getRoomCapacity(hotel)}
-                  </span>
-                </HotelsWrapper>
-              ))
-            ) : (
-              <></>
-            )}
-          </div>
-        </SectionWrapper>
-      )}
+      {isValidTicket(ticket) && hotels && <Hotels hotels={hotels} />}
     </>
   );
 }
